requirements.txt
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
pyserial==3.5
pytest==8.3.5
